Introduce SOCI to persiste in sqlite3

+++ backend/include/myroomies/utils/db/Def.h
@@ -6,7 +6,7 @@ namespace myroomies {
 namespace utils {
 namespace db {
 
-typedef uint32_t Key_t;
+typedef uint64_t Key_t;
 typedef Key_t ForeignKey_t;
 
 } /* namespace db */
